master_flow/pyproject.toml
[project]
name = "master_flow"
version = "0.1.0"
description = "master-flow using crewAI"
authors = [{ name = "Ivan", email = "ivantham789@gmail.com" }]
requires-python = ">=3.10,<3.14"
dependencies = [
    "crewai[tools,google-genai]==1.9.3",
    "fastapi",
    "uvicorn",
    "qdrant-client",
    "sentence-transformers",
    "ddgs",
    "python-dotenv"
]

[project.scripts]
kickoff = "master_flow.main:kickoff"
run_crew = "master_flow.main:kickoff"
plot = "master_flow.main:plot"
run_with_trigger = "master_flow.main:run_with_trigger"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "flow"
